Added auto discovery and exposed inputs as buttons
The is an extended version of Steve Vibert's (SteveV) driver files of the same name.  If you have found this through a search, this is and Alpha version.  Steve will release the more fully tested version.

This requires the child driver located here.  It will not work with the current version Steve has published.